requirements.txt
python-dotenv==1.0.1
python-json-logger==3.3.0
SQLAlchemy==2.0.39
asyncpg==0.30.0
aiohttp==3.11.14
backoff==2.2.1
ollama==0.4.7